requirements.txt
tensorflow
numpy
pandas
dvc
notebook
matplotlib
seaborn
python-box==6.0.2
pyYaml
tdqm
ensure ==1.0.2
joblib
types-pyYaml
scipy
Flask
Flask-cors
-e .